requirements.txt
asttokens==3.0.0
attrs==25.3.0
awkward==1.10.5
awkward_cpp==40
backcall==0.2.0
cachetools==5.5.2
comm==0.2.2
contourpy==1.1.1
cramjam==2.10.0
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
executing==2.2.0
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
h5py==3.11.0
hepunits==2.3.6
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
-e git+https://github.com/rmastand/madminer.git@70b28308ce795be04d83a6768e438e44d17124d2#egg=madminer
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.23.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==25.0
parso==0.8.4
particle==0.25.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==27.0.0
SciencePlots==2.1.1
scipy==1.10.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
torch==2.4.1
tornado==6.4.2
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.13.2
uproot==4.3.7
uproot-methods==0.8.0
vector==1.6.3
wcwidth==0.2.13
xxhash==3.5.0
zipp==3.20.2
